+++ momoim-external-api/src/main/java/com/triplem/momoim/api/review/service/ReviewService.java
@@ -1,4 +1,4 @@
-package com.triplem.momoim.api.review;
+package com.triplem.momoim.api.review.service;
 
 import com.triplem.momoim.core.common.PaginationInformation;
 import com.triplem.momoim.core.domain.review.ModifyReview;
